backend/pyproject.toml
[tool.poetry]
name = "caster_dashboard_2"
version = "2.0.0-beta1"
description = ""
authors = ["Thorsten Schmitt <dev@tschmitt.eu>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-filter = "^21.1"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
channels = "^3.0.4"
channels-redis = "^3.3.1"
argon2-cffi = "^21.3.0"
autopep8 = "^1.6.0"
Pillow = "^8.4.0"
pylint = "^2.12.2"
psycopg2-binary = "^2.9.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
